backend/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements/requirements.in
#
asgiref==3.7.2
    # via
    #   channels
    #   channels-redis
    #   django
channels==4.0.0
    # via
    #   -r requirements/requirements.in
    #   channels-redis
channels-redis==4.2.0
    # via -r requirements/requirements.in
configparser==6.0.0
    # via -r requirements/requirements.in
django==4.2.7
    # via
    #   -r requirements/requirements.in
    #   channels
    #   django-extensions
    #   djangorestframework
    #   djangorestframework-simplejwt
django-extensions==3.2.3
    # via -r requirements/requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements/requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.3.1
    # via -r requirements/requirements.in
msgpack==1.0.7
    # via channels-redis
pillow==10.1.0
    # via -r requirements/requirements.in
psycopg2==2.9.9
    # via -r requirements/requirements.in
pyjwt==2.8.0
    # via djangorestframework-simplejwt
pytz==2023.3.post1
    # via djangorestframework
redis==5.0.1
    # via channels-redis
sqlparse==0.4.4
    # via django
tzdata==2023.3
    # via django
